requirements.txt
bitarray==1.2.1
Django==3.1
eyeD3==0.9.5
ffmpeg==1.4
glob2==0.7
HeapDict==1.0.1
hmmlearn==0.2.3
imageio==2.6.1
imageio-ffmpeg==0.4.2
imagesize==1.2.0
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.1
jsonschema==3.2.0
matplotlib==3.1.3
nltk==3.4.5
numpy==1.18.1
pandas==1.0.1
path==13.1.0
pathlib2==2.3.5
pathtools==0.1.2
Pillow==7.0.0
plotly==4.9.0
pyAudioAnalysis==0.3.5
pydub==0.24.1
requests==2.22.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
SoundFile==0.10.3.post1
SpeechRecognition==3.8.1
SQLAlchemy==1.3.13
sqlparse==0.3.1
statsmodels==0.11.0
tqdm==4.42.1
azure-cognitiveservices-speech==1.13.0